requirements.txt
discord.py
dnspython==2.0.0
PyNaCl==1.3.0
async-timeout==3.0.1
beautifulsoup4==4.9.1
matplotlib==3.3.2
lxml
requests
flask